Mercury to main (#58)

+++ tests/helpers.py
@@ -67,7 +67,8 @@ def delete_existing_file(file):
 
 
 def delete_existing_folder(folder):
-    shutil.rmtree(folder)
+    if os.path.exists(folder):
+        shutil.rmtree(folder)
 
 
 def close_logger(logger):
@@ -128,7 +129,7 @@ class BuildersCommon:
     @staticmethod
     def demographics_builder(aliens_distribution=None, total_population=500):
         from emodpy.demographics.demographics import Demographics
-        from emod_api.demographics.Node import Node
+        from emod_api.demographics.node import Node
 
         if not aliens_distribution:
             aliens_distribution = [0.3, 0.3, 0.4]
